move autolayout graph into graph
