+++ src/test.py
@@ -1,6 +1,5 @@
 from manim import *
 from src.graph import FlowGraph
-from src.auto_layout_graph import getEdgesAndVerticesAsMobjects
 from src.vertices_examples import VerticesExamples as V
 
 
@@ -16,10 +15,7 @@ class Test(Scene):
             5: [2, 0, 0],
         }
 
-        vertices, edges = getEdgesAndVerticesAsMobjects(
-            vertices, edges, capacities, layout=lt
-        )
-        graph = FlowGraph(vertices, edges)
+        graph = FlowGraph(vertices, edges, capacities, layout=lt)
         self.camera.background_color = WHITE
 
         self.add(graph)
@@ -33,18 +29,16 @@ class Test2(Scene):
             0: [0, -1, 0],
             1: [2, 1, 0],
         }
-        vertices, edges = getEdgesAndVerticesAsMobjects(
-            vertices, edges, capacities, layout=lt
-        )
 
-        graph = FlowGraph(vertices, edges)
+        graph = FlowGraph(vertices, edges, capacities, layout=lt)
         self.camera.background_color = WHITE
 
         self.add(graph)
 
-        edges[0].add_to_current_flow(20, self)
+        """ edges[0].add_to_current_flow(20, self)
         self.wait(3, frozen_frame=False)
         edges[0].add_to_current_flow(50, self)
         self.wait(10, frozen_frame=False)
         edges[0].add_to_current_flow(20, self)
         self.wait(3, frozen_frame=False)
+ """
